Add some testing tooling
A simple container file based on Alpine Edge, and a simple Makefile for
building and running the container.  test.sh can be used for logging in
and sending GET requests.  Extending with POST should be simple but it's
not currently supported.

malicek.pl now packages all these files, plus README.

README now references mlck and is reformatted with tw=72.

Signed-off-by: Petr Šabata <contyk@redhat.com>