More cleanup
Create classes for Lednička and convert the code to it.  Rewrite it to
use strict regexes.

Rewrite parse_rooms() regexes to be strict and optimized.  Add support
for admin class detection, too.

Stop sorting users and rooms in parse_rooms(), make the clients do it,
if they wish.  They now get it in the order we get it.

Signed-off-by: Petr Šabata <contyk@redhat.com>