Major breaking changes in many things
Dropping the /api resource.  All public files are now server statically,
automatically.  This means / is no longer a status page, that was moved
to /malicek.

It also means one no longer needs a special forward proxy to use the
bundled web interface.  It's all accessible via the same path.

Web interface adapted to these changes, as well as other API changes.
This is useful for testing, even though I'm planning on replacing it
with something entirely different.

The web interface no longer serves the spinner, simplifying licensing.

I don't intend to maintain cli.pl, so that's dropped.  Use the web
interface or mlck.

parse_rooms() now handles rooms with many people (shown with different
layout on Alik), as well as permanent but empty rooms with fake topics.

It should now be possible to create new rooms by POSTing "name" to
/rooms, although that's entirely untested.  Returns 409 on failure or
redirects to the room resource with 201 on success.

Signed-off-by: Petr Šabata <contyk@redhat.com>