Deal with the new default avatars
Unlike all the other avatars, default avatars are directly on Alik.
Now, the new Alik link actually takes form of /-/avatar/link and
redirects to o00o.cz if an avatar is defined, but we do not know the
link in chat messages, so we cannot rely on that to simplify this
further.

Signed-off-by: Petr Šabata <contyk@redhat.com>