Room list updates
Two changes here.

First, don't ignore rooms that have been quiet for way too long.  That's
been a long-standing bug.

Second, the recent changes now list room admins instead of creators in
some cases.  This happens when it's a permanent room and the admin is
not currently present.  If it's a permanent room and the moderator is
present, nothing is shown.  If it's a temporary room, the room creator
is shown, regardless of whether they're the moderator or the moderator's
presence.

We should probably capture the moderator and expose them over the API
but that is not part of this change.

Signed-off-by: Petr Šabata <contyk@contyk.dev>